Feat: add path aliases support
