+++ src/App.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom';
-import { RegisterForm } from './components/RegisterForm';
-import { LoginForm } from './components/LoginForm';
-import { Dashboard } from './components/Dashboard';
+import { RegisterForm, LoginForm } from '@components/auth';
+import { Dashboard } from '@components/Dashboard';
 
 function App() {
   return (
