+++ src/components/Dashboard.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Sidebar } from './Sidebar';
 import { TopBar } from './TopBar';
-import { useTheme } from '../hooks/useTheme';
-import { Project } from '../types/monitoring';
+import { useTheme } from '@hooks';
+import { Project } from '@types';
 import { MonitoringDashboard } from './monitoring/MonitoringDashboard';
 import { AutomatedTesting } from './testing/AutomatedTesting';
 import { CICDPipeline } from './pipelining/CICDPipeline';
