+++ src/components/auth/LoginForm.tsx
@@ -1,15 +1,13 @@
 import React, { useState } from 'react';
 import { useNavigate, Link } from 'react-router-dom';
-import { UserPlus } from 'lucide-react';
-import { User, db } from '../db/db';
-import { useTheme } from '../hooks/useTheme';
+import { LogIn } from 'lucide-react';
+import { db } from '@/db/db';
+import { useTheme } from '@hooks';
 
-export function RegisterForm() {
+export default function LoginForm() {
   const navigate = useNavigate();
   const { darkMode, setDarkMode } = useTheme();
-  const [formData, setFormData] = useState<Omit<User, 'id'>>({
-    businessName: '',
-    email: '',
+  const [formData, setFormData] = useState({
     username: '',
     password: ''
   });
@@ -18,15 +16,19 @@ export function RegisterForm() {
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
     try {
-      const existingUser = await db.users.where('username').equals(formData.username).first();
-      if (existingUser) {
-        setError('Username already exists');
-        return;
+      const user = await db.users
+        .where('username')
+        .equals(formData.username)
+        .first();
+      
+      if (user && user.password === formData.password) {
+        localStorage.setItem('currentUser', JSON.stringify(user));
+        navigate('/dashboard');
+      } else {
+        setError('Invalid username or password');
       }
-      await db.users.add(formData);
-      navigate('/login');
     } catch (err) {
-      setError('Registration failed');
+      setError('Login failed');
     }
   };
 
@@ -34,10 +36,10 @@ export function RegisterForm() {
     <div className="min-h-screen bg-gray-50 dark:bg-gray-900 flex flex-col justify-center py-12 sm:px-6 lg:px-8">
       <div className="sm:mx-auto sm:w-full sm:max-w-md">
         <div className="flex justify-center">
-          <UserPlus className="h-12 w-12 text-indigo-600 dark:text-indigo-400" />
+          <LogIn className="h-12 w-12 text-indigo-600 dark:text-indigo-400" />
         </div>
         <h2 className="mt-6 text-center text-3xl font-extrabold text-gray-900 dark:text-gray-100">
-          Register your business
+          Sign in to your account
         </h2>
         <div className="mt-2 flex justify-center">
           <button
@@ -52,40 +54,6 @@ export function RegisterForm() {
       <div className="mt-8 sm:mx-auto sm:w-full sm:max-w-md">
         <div className="bg-white dark:bg-gray-800 py-8 px-4 shadow sm:rounded-lg sm:px-10">
           <form className="space-y-6" onSubmit={handleSubmit}>
-            <div>
-              <label htmlFor="businessName" className="block text-sm font-medium text-gray-700 dark:text-gray-300">
-                Business Name
-              </label>
-              <div className="mt-1">
-                <input
-                  id="businessName"
-                  name="businessName"
-                  type="text"
-                  required
-                  className="appearance-none block w-full px-3 py-2 border border-gray-300 dark:border-gray-600 rounded-md shadow-sm placeholder-gray-400 dark:placeholder-gray-500 bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500"
-                  value={formData.businessName}
-                  onChange={(e) => setFormData({ ...formData, businessName: e.target.value })}
-                />
-              </div>
-            </div>
-
-            <div>
-              <label htmlFor="email" className="block text-sm font-medium text-gray-700 dark:text-gray-300">
-                Email
-              </label>
-              <div className="mt-1">
-                <input
-                  id="email"
-                  name="email"
-                  type="email"
-                  required
-                  className="appearance-none block w-full px-3 py-2 border border-gray-300 dark:border-gray-600 rounded-md shadow-sm placeholder-gray-400 dark:placeholder-gray-500 bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500"
-                  value={formData.email}
-                  onChange={(e) => setFormData({ ...formData, email: e.target.value })}
-                />
-              </div>
-            </div>
-
             <div>
               <label htmlFor="username" className="block text-sm font-medium text-gray-700 dark:text-gray-300">
                 Username
@@ -129,13 +97,13 @@ export function RegisterForm() {
                 type="submit"
                 className="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-indigo-600 hover:bg-indigo-700 dark:bg-indigo-500 dark:hover:bg-indigo-600 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 dark:focus:ring-offset-gray-900"
               >
-                Register
+                Sign in
               </button>
             </div>
 
             <div className="text-center mt-4">
-              <Link to="/login" className="text-indigo-600 dark:text-indigo-400 hover:text-indigo-500 dark:hover:text-indigo-300">
-                Already have an account? Sign in!
+              <Link to="/register" className="text-indigo-600 dark:text-indigo-400 hover:text-indigo-500 dark:hover:text-indigo-300">
+                Don't have an account? Sign up!
               </Link>
             </div>
           </form>
