+++ src/components/auth/RegisterForm.tsx
@@ -1,13 +1,15 @@
 import React, { useState } from 'react';
 import { useNavigate, Link } from 'react-router-dom';
-import { LogIn } from 'lucide-react';
-import { db } from '../db/db';
-import { useTheme } from '../hooks/useTheme';
+import { UserPlus } from 'lucide-react';
+import { User, db } from '@/db/db';
+import { useTheme } from '@hooks';
 
-export function LoginForm() {
+export default function RegisterForm() {
   const navigate = useNavigate();
   const { darkMode, setDarkMode } = useTheme();
-  const [formData, setFormData] = useState({
+  const [formData, setFormData] = useState<Omit<User, 'id'>>({
+    businessName: '',
+    email: '',
     username: '',
     password: ''
   });
@@ -16,19 +18,15 @@ export function LoginForm() {
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
     try {
-      const user = await db.users
-        .where('username')
-        .equals(formData.username)
-        .first();
-      
-      if (user && user.password === formData.password) {
-        localStorage.setItem('currentUser', JSON.stringify(user));
-        navigate('/dashboard');
-      } else {
-        setError('Invalid username or password');
+      const existingUser = await db.users.where('username').equals(formData.username).first();
+      if (existingUser) {
+        setError('Username already exists');
+        return;
       }
+      await db.users.add(formData);
+      navigate('/login');
     } catch (err) {
-      setError('Login failed');
+      setError('Registration failed');
     }
   };
 
@@ -36,10 +34,10 @@ export function LoginForm() {
     <div className="min-h-screen bg-gray-50 dark:bg-gray-900 flex flex-col justify-center py-12 sm:px-6 lg:px-8">
       <div className="sm:mx-auto sm:w-full sm:max-w-md">
         <div className="flex justify-center">
-          <LogIn className="h-12 w-12 text-indigo-600 dark:text-indigo-400" />
+          <UserPlus className="h-12 w-12 text-indigo-600 dark:text-indigo-400" />
         </div>
         <h2 className="mt-6 text-center text-3xl font-extrabold text-gray-900 dark:text-gray-100">
-          Sign in to your account
+          Register your business
         </h2>
         <div className="mt-2 flex justify-center">
           <button
@@ -54,6 +52,40 @@ export function LoginForm() {
       <div className="mt-8 sm:mx-auto sm:w-full sm:max-w-md">
         <div className="bg-white dark:bg-gray-800 py-8 px-4 shadow sm:rounded-lg sm:px-10">
           <form className="space-y-6" onSubmit={handleSubmit}>
+            <div>
+              <label htmlFor="businessName" className="block text-sm font-medium text-gray-700 dark:text-gray-300">
+                Business Name
+              </label>
+              <div className="mt-1">
+                <input
+                  id="businessName"
+                  name="businessName"
+                  type="text"
+                  required
+                  className="appearance-none block w-full px-3 py-2 border border-gray-300 dark:border-gray-600 rounded-md shadow-sm placeholder-gray-400 dark:placeholder-gray-500 bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500"
+                  value={formData.businessName}
+                  onChange={(e) => setFormData({ ...formData, businessName: e.target.value })}
+                />
+              </div>
+            </div>
+
+            <div>
+              <label htmlFor="email" className="block text-sm font-medium text-gray-700 dark:text-gray-300">
+                Email
+              </label>
+              <div className="mt-1">
+                <input
+                  id="email"
+                  name="email"
+                  type="email"
+                  required
+                  className="appearance-none block w-full px-3 py-2 border border-gray-300 dark:border-gray-600 rounded-md shadow-sm placeholder-gray-400 dark:placeholder-gray-500 bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500"
+                  value={formData.email}
+                  onChange={(e) => setFormData({ ...formData, email: e.target.value })}
+                />
+              </div>
+            </div>
+
             <div>
               <label htmlFor="username" className="block text-sm font-medium text-gray-700 dark:text-gray-300">
                 Username
@@ -97,13 +129,13 @@ export function LoginForm() {
                 type="submit"
                 className="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-indigo-600 hover:bg-indigo-700 dark:bg-indigo-500 dark:hover:bg-indigo-600 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 dark:focus:ring-offset-gray-900"
               >
-                Sign in
+                Register
               </button>
             </div>
 
             <div className="text-center mt-4">
-              <Link to="/register" className="text-indigo-600 dark:text-indigo-400 hover:text-indigo-500 dark:hover:text-indigo-300">
-                Don't have an account? Sign up!
+              <Link to="/login" className="text-indigo-600 dark:text-indigo-400 hover:text-indigo-500 dark:hover:text-indigo-300">
+                Already have an account? Sign in!
               </Link>
             </div>
           </form>
