+++ src/components/auth/index.ts
@@ -1,0 +1,2 @@
+export { default as RegisterForm } from './RegisterForm';
+export { default as LoginForm } from './LoginForm';
