+++ src/components/shared/index.ts
@@ -1,0 +1,2 @@
+export { default as BoxedWrapper } from './BoxedWrapper';
+export { default as SectionHeader } from './SectionHeader';
